+++ src/Designer/backend/src/Designer/Services/Implementation/UiFoldersService.cs
@@ -410,7 +410,7 @@ public class UiFoldersService : IUiFoldersService
                         {
                             ["size"] = "L",
                             ["id"] = "service-task-title",
-                            ["type"] = "Header",
+                            ["type"] = "Heading",
                             ["textResourceBindings"] = new JsonObject
                             {
                                 ["title"] = "service_task_custom_pdf_default.title",